- Added Cap attempts for quests. This will make it only use the minimum stamina for the quest in question. So if I set Jade Mining to 15 max it will always go for using 15 stamina even if heroes have more stamina when ready. - Added Quest Instantly. Quest Instantly sends heroes on quests immediately when they're ready without waiting for any other heroes, or the whole group if there are more ready. - Added tooltips to Bot Settings Priority Fee settings, Quest Eagerly, Quest Instantly and Cap Attempts to explain what they do. - Fixed a bug where stamina potions was not being used if a hero had more than 5 stamina. It is now correctly set to max stam - 20. - Added logs to tell you if Quest Instantly and Cap Attempts are enabled when it sends heroes on quests. - Added decimals to priority fee bot settings. - Fixed a bug that would crash the bot upon restart. - Fixed a bug that would crash the bot if heroes didn't get a suggested quest.

+++ Bot/DFKBotSettings.cs
@@ -15,12 +15,24 @@ namespace PirateQuester.Bot
 				new()
 				{
 					Chain = Constants.ChainsList[0],
-					QuestEnabled = Enumerable.Range(0, 25).Select(i => new QuestEnabled() { Enabled = true, QuestId = i, QuestEagerly = eagerQuests.Contains(i) }).ToList()
+					QuestEnabled = Enumerable.Range(0, 25).Select(i => 
+					new QuestEnabled() 
+					{ 
+						Enabled = true, 
+						QuestId = i, 
+						QuestEagerly = eagerQuests.Contains(i)
+					}).ToList()
 				},
 				new ()
 				{
 					Chain = Constants.ChainsList[1],
-					QuestEnabled = Enumerable.Range(0, 23).Select(i => new QuestEnabled() { Enabled = true, QuestId = i, QuestEagerly = eagerQuests.Contains(i) }).ToList()
+					QuestEnabled = Enumerable.Range(0, 23).Select(i => 
+					new QuestEnabled() 
+					{ 
+						Enabled = true, 
+						QuestId = i, 
+						QuestEagerly = eagerQuests.Contains(i) 
+					}).ToList()
 				}
 			};
 		}
@@ -29,10 +41,10 @@ namespace PirateQuester.Bot
 		public int CancelTxnDelay { get; set; } = 60000;
 		public int UpdateInterval { get; set; } = 180;
 		public int MinStamina { get; set; } = 20;
-        public int MaxGasFeeGwei { get; set; } = 200;
-        public int MaxGasFeeGweiKlaytn { get; set; } = 40;
-        public int MaxPriorityFeeGwei { get; set; } = 0;
-        public int MaxPriorityFeeGweiKlaytn { get; set; } = 0;
+        public decimal MaxGasFeeGwei { get; set; } = 200;
+        public decimal MaxGasFeeGweiKlaytn { get; set; } = 40;
+        public decimal MaxPriorityFeeGwei { get; set; } = 0;
+        public decimal MaxPriorityFeeGweiKlaytn { get; set; } = 0;
         public decimal WarnFloorPercentage { get; set; } = 90;
 		public bool LevelUp { get; set; } = true;
 		public bool UseStaminaPotions { get; set; } = true;
