+++ Bot/DFKBotSettingsDTO.cs
@@ -12,10 +12,10 @@ namespace PirateQuester.Bot
 		public int CancelTxnDelay { get; set; }
 		public int UpdateInterval { get; set; }
 		public int MinStamina { get; set; }
-		public int MaxGasFeeGwei { get; set; }
-        public int MaxGasFeeGweiKlaytn { get; set; }
-        public int MaxPriorityFeeGwei { get; set; }
-        public int MaxPriorityFeeGweiKlaytn { get; set; }
+		public decimal MaxGasFeeGwei { get; set; }
+        public decimal MaxGasFeeGweiKlaytn { get; set; }
+        public decimal MaxPriorityFeeGwei { get; set; }
+        public decimal MaxPriorityFeeGweiKlaytn { get; set; }
         public bool LevelUp { get; set; }
 		public bool ForceStampotOnFullXP { get; set; }
 		public bool UseStaminaPotions { get; set; }
